% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_enrichment.R
\name{custard}
\alias{custard}
\title{Test enrichment of profiles in short time course gene expression data.}
\usage{
custard(mat, templates = 50, magnitude = 1, alpha = 0.05,
  permutations = 1000, mc.cores = 4)
}
\arguments{
\item{mat}{A matrix with gene expression profiles (columns).}

\item{templates}{An integer, or else a matrix of templates.}

\item{magnitude}{At each time point, a gene may change up to this many units.
Only used if \code{templates} is an integer.}

\item{alpha}{Significance level (uncorrected for multiple testing).
Default is 0.05.}

\item{permutations}{Compute this many permutations of the time points.}

\item{mc.cores}{Use this many cores to compute the permutations.}
}
\value{
A list with results.
}
\description{
Test enrichment of profiles in short time course gene expression data.
}

