% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{distinct_templates}
\alias{distinct_templates}
\title{Make all possible templates for short time series gene expression data.}
\usage{
distinct_templates(all_templates, templates = 10)
}
\arguments{
\item{all_templates}{A matrix, as returned by \code{\link{make_templates}}}

\item{templates}{An integer.}
}
\value{
A matrix with one row per time point and one column per template.
}
\description{
Make all possible templates for short time series gene expression data.
}

