% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{make_templates}
\alias{make_templates}
\title{Make all possible templates for short time series gene expression data.}
\usage{
make_templates(timepoints = 3, magnitude = 1)
}
\arguments{
\item{timepoints}{An integer between 3 and 9.}

\item{magnitude}{An integer between 1 and 4.}
}
\value{
A matrix with one row per time point and one column per template.
}
\description{
Make all possible templates for short time series gene expression data.
}

